Refactored DataStore -> Relocated to source + Pydantic models with validations + module imports for module + updated Writers (including tests) + added custom tests
